sedispatch: Use timeout for reading data from audispd
auparse_feed() can't recognize if an event is complete when the event
contains only one AVC record. In this case, a message about new AVC is
resend after next audit event come which could be several minutes lates.

With this patch, we expect that the event is complete when nothing has
come in last 3 seconds.

Fixes:
https://bugzilla.redhat.com/show_bug.cgi?id=1322771
https://github.com/fedora-selinux/setroubleshoot/issues/13