plugins: fix restorecon "then" text
In case a syscall fails due to insufficient permissions to parent
directory of target object, PATH record provided by audit will contain
path to the target object of the syscall (not the parent directory).
Unfortunately I haven't found a way to distinguish this case (let alone
determine the parent directory causing the denial).

Update restorecon plugin "then" text to let the user know that the
provided command may not help (since restorecon is called on the wrong
object).

fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1332097

Signed-off-by: Vit Mojzis <vmojzis@redhat.com>