framework: Increase the space for suggested solutions in sealert
After clicking the 'Troubleshoot' button, the user had to manually enlarge the
window size to see the solutions suggested by the plugins.

This patch increases the original window size to increase the amouth of
information displayed to the user. Generally, all of the information is shown,
even for larger texts.

Environment used for the test
- setroubleshoot-3.3.13
- setroubleshoot-plugins-3.3.8
- Fedora 26