framework: Highlight suggestions with the highest confidence
After clicking the 'Troubleshoot' button, the texts and buttons ('Plugin
Details', 'Report Bug') for all suggestions are greyed out by default. This is
confusing to the user, making her believe the buttons are inaccessible, while
in fact they are accessible after clicking the corresponding 'If you want...'
button on the left. However, this requirement is in no way obvious to the user.

This fix greys out all suggestions EXCEPT the ones with the highest confidence.
Note more suggestions can be highlighted.