plugins/bind_ports: Do not use when there are no allowed_target_types
Do not use this plugin when source domain is not allowed to name_bind
to any port type. This caused invalid command suggestion such as:

semanage port -a -t  -p udp 5382 (the type is missing)

This situation would require the user to add an allow rule or switch a
boolean, which is handled by the catchall and catchall_boolean plugins.

Fixes: rhbz#1255627