framework: Cancel pending alarm during AVC analyses
Previously the alarm was reset only in main DBUS thread and only when a new AVC
appeared. In cases when there were several AVC messages in short time, analyses
could take more than a default timeout and later analyses were not saved to
the database. Now we cancel pending timeouts before analyze_avc() and reset the
timeout back to default when it's done.

Fixes:
    $ journalctl | grep 'sealert -l'
    setroubleshoot[314039]: SELinux is preventing bash from search access on the directory .local. For complete SELinux messages run: sealert -l ccf3307a-f4ab-4584-87c6-63884daf841a

    $ sealert -l ccf3307a-f4ab-4584-87c6-63884daf841a
    Error
    query_alerts error (1003): id (ccf3307a-f4ab-4584-87c6-63884daf841a) not found