html_util: remove html_document
I suspect this can safely be removed, because I don't think it
has worked for six years. Since dbaaf0f it uses `six.string_types`
but does not `import six`; if the function were ever called, it
would crash immediately. I suggest this is reasonable proof that
it's useless. Nothing in setroubleshoot itself calls it.

Signed-off-by: Adam Williamson <awilliam@redhat.com>